--- experimental/reporting-ui/src/main/react/reporting-ui/component/chart/chart_helper/BUILD.bazel
@@ -1,18 +1,0 @@
-load("@aspect_rules_js//js:defs.bzl", "js_library")
-
-package(
-    default_visibility = [
-        "//experimental/reporting-ui/src/main/react/reporting-ui:__subpackages__",
-    ],
-)
-
-js_library(
-    name = "chart_helper",
-    srcs = [
-        "chart_factory.tsx",
-        "helpers.js",
-    ],
-    deps = [
-        "//experimental/reporting-ui/src/main/react/reporting-ui/public/asset/icon",
-    ],
-)
