+++ experimental/reporting-ui/src/main/react/reporting-ui/component/chart/chart_type/index.tsx
