+++ experimental/reporting-ui/src/main/react/reporting-ui/component/chart/chart_type/on_target_reach.tsx
@@ -13,36 +13,26 @@
 // limitations under the License.
 
 import React from 'react';
-import { ChartFactory, ChartType } from './chart_helper/chart_factory';
-import { TargetReach } from '../../model/reporting';
+import { ChartFactory, ChartType } from '../chart_factory';
+import { Reach } from '../../../model/reporting';
 
-const neutralColors = [
-    "#CACACA",
-    "#959595",
-    "#5E5E5E",
-]
-
-type OnTargetXmediaReachByFreqProps = {
+type OnTargetReachProps = {
     id: string,
-    reach: TargetReach[],
+    reach: Reach[],
+    pubColors: { [Name: string]: string}
 }
 
-export function OnTargetXmediaReachByFreq({id, reach}: OnTargetXmediaReachByFreqProps) {
+export function OnTargetReach({id, reach, pubColors}: OnTargetReachProps) {
     const config = {
-        catColors: {}
+        pubColors,
     }
-    const colors = {}
-    const unique = [...new Set(reach?.map(item => item.cat))];
-    unique.forEach((x: any, i) => colors[x] = neutralColors[i])
-    config.catColors = colors;
-
     return (
         <ChartFactory
             cardId={id}
-            title='On target cross-media reach by frequency'
+            title='On target reach'
             data={reach}
             config={config}
-            type={ChartType.percentMultiLine}
+            type={ChartType.multiLine}
         />
     )
 }
