+++ experimental/reporting-ui/src/main/react/reporting-ui/view/report/component/BUILD.bazel
@@ -23,7 +23,7 @@ js_library(
     ],
     deps = [
         "//experimental/reporting-ui/src/main/react/reporting-ui/component/card_wrapper",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/component/chart",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/component/chart/chart_type",
         "//experimental/reporting-ui/src/main/react/reporting-ui/public/asset/icon",
         "//experimental/reporting-ui/src/main/react/reporting-ui/util:formatting",
     ],
